Copper buttons + misc things
The access transformer is so the method that sets the pulse length on buttons is public.

The change in GenerationEvents is just a small tweak to make it look slightly cleaner and to better utilize blueprint.

There might be the odd unused import here or there but I will fix them later.