Brazilian Portuguese Translation
As promised, here is the PT-BR translation.
There were two tortosie strings equivalent to vanilla ones that are mistranslated in vanilla, with one of them being swapped with the other and said other just being copied from another close string. I chose to correct this in this translation over being consistent with the mistake, swapping them back and creating a new string for the one which is missing one in vanilla. This does mean that if vanilla ever corrects this there is a chance this will be inconsistent, and the chances I'll notice the correction are very small, but this is so minor I don't think it is a big problem.